Integrated w/ Google Maps API
Now capable of taking in any string input reresenting a location, doing what I'm assuming is analogous to a Google Maps search with the string, and then finding the coordinates corresponding to the given location.

Also outputs list of bus stops near the location, but that's a vestige of prior development activity.